FuriCore: move type casting to dangerous_defines
