+++ core/furi.h
@@ -7,6 +7,7 @@
 #include <cmsis_os2.h>
 
 #include <furi/common_defines.h>
+#include <furi/dangerous_defines.h>
 #include <furi/check.h>
 #include <furi/memmgr.h>
 #include <furi/memmgr_heap.h>
