+++ core/furi/common_defines.h
@@ -81,27 +81,3 @@
 #ifndef FURI_CRITICAL_EXIT
 #define FURI_CRITICAL_EXIT() __set_PRIMASK(primask_bit)
 #endif
-
-#ifndef FURI_CONST_ASSIGN
-#define FURI_CONST_ASSIGN_(T, x, y) \
-    ({                              \
-        T* tmp_x = (T*)&x;          \
-        *tmp_x = y;                 \
-        *tmp_x;                     \
-    })
-#define FURI_CONST_ASSIGN(x, y)                                        \
-    _Generic((x), signed char                                          \
-             : FURI_CONST_ASSIGN_(signed char, x, y), unsigned char    \
-             : FURI_CONST_ASSIGN_(unsigned char, x, y), short          \
-             : FURI_CONST_ASSIGN_(short, x, y), unsigned short         \
-             : FURI_CONST_ASSIGN_(unsigned short, x, y), int           \
-             : FURI_CONST_ASSIGN_(int, x, y), unsigned                 \
-             : FURI_CONST_ASSIGN_(unsigned, x, y), long                \
-             : FURI_CONST_ASSIGN_(long, x, y), unsigned long           \
-             : FURI_CONST_ASSIGN_(unsigned long, x, y), long long      \
-             : FURI_CONST_ASSIGN_(long long, x, y), unsigned long long \
-             : FURI_CONST_ASSIGN_(unsigned long long, x, y), float     \
-             : FURI_CONST_ASSIGN_(float, x, y), double                 \
-             : FURI_CONST_ASSIGN_(double, x, y), long double           \
-             : FURI_CONST_ASSIGN_(long double, x, y))
-#endif
